Move validation methods to extensions methods
